requirements.txt
python-telegram-bot[job-queue]==21.5
python-dotenv==1.0.1
google-generativeai==0.8.3
gspread==6.1.2
google-auth==2.27.0
apscheduler==3.10.4
pytz==2024.1
